requirements.txt
fastmcp
apscheduler
sqlalchemy
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
uvicorn
fastapi
python-dotenv
cryptography